requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
clean-text==0.6.0
click==8.1.3
cloudevents==1.9.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.6
emoji==1.7.0
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.4
ftfy==6.1.1
functions-framework==3.4.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
imageai==3.0.3
ipykernel==6.23.1
ipython==8.14.0
itsdangerous==2.1.2
jax==0.4.11
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
ndjson==0.3.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
pyspellchecker==0.7.2
pythainlp==4.0.2
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
watchdog==3.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
xlrd==2.0.1